+++ interfaces/console/src/components/NodeStatus/index.tsx
@@ -53,7 +53,7 @@ const NodeStatus = ({
 
     return (
         <Grid container>
-            <Grid item xs={12} md={9}>
+            <Grid item xs={12} md={8}>
                 <NodeDropDown
                     nodes={nodes}
                     loading={loading}
@@ -66,7 +66,7 @@ const NodeStatus = ({
             </Grid>
             <Grid
                 item
-                md={3}
+                md={4}
                 xs={12}
                 container
                 spacing={2}
